chore: update packages
