+++ packages/date-fns/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@adamdotai/date-fns",
-  "version": "3.0.0",
+  "version": "3.1.1",
   "description": "Abstraction over common javascript date management libraries",
   "main": "build/index.js",
   "module": "build/index.esm.js",
